changes in socket files and readme
