+++ README.md
@@ -54,6 +54,30 @@ try:
 except Exception as e:
     print("Logout failed: {}".format(e.message))
 
+#websocket
+from smartapi.smartSocket import SmartSocket
+FEED_TOKEN= "your feed token"
+CLIENT_CODE="your client code"
+token="channel you want the information of" #"nse_cm|2885&nse_cm|1594&nse_cm|11536"
+
+ss = SmartSocket(FEED_TOKEN, CLIENT_CODE)
+
+def on_tick(ws, tick):
+    print("Ticks: {}".format(tick))
+
+def on_connect(ws, response):
+    ws.send_request(token)
+
+def on_close(ws, code, reason):
+    ws.stop()
+
+# Assign the callbacks.
+ss.on_ticks = on_tick
+ss.on_connect = on_connect
+ss.on_close = on_close
+
+ss.connect( )
+
 
 License
 MIT
